fix: github stats in different file
