+++ frontend/src/lib/utils/update-spots-occupied.ts
